Re-organized code to group most of the clicast commands into a new file
