+++ src/testPane.ts
@@ -15,6 +15,7 @@ import {
 } from "vscode";
 
 
+import { clicastCommandToUse } from "./clicast";
 import { updateDisplayedCoverage, updateCOVdecorations } from "./coverage";
 
 import {
@@ -33,25 +34,21 @@ import {
   getFunctionNameFromID,
   getEnviroNameFromID,
   getTestNameFromID,
-  getTestNode,
   getUnitNameFromID,
   testNodeType,
 } from "./testData";
 
 import {
-  clicastCommandToUse,
   executeCommand,
   loadLaunchFile,
   addLaunchConfiguration,
 } from "./utilities";
 import {
   cfgOptionType,
-  getEnviroNameFromScript,
   getVcastOptionValues,
 } from "../src-common/commonUtilities";
 
 import {
-  getClicastArgsFromTestNode,
   getEnviroDataFromPython,
   getResultFileForTest,
   globalTestStatusArray,
@@ -63,7 +60,6 @@ import {
 } from "./vcastTestInterface";
 
 const fs = require("fs");
-const os = require("os");
 const path = require("path");
 
 // find test location in file
@@ -135,134 +131,6 @@ function addTestNodes(
   }
 }
 
-function convertTestScriptContents(scriptPath: string) {
-  // Read the file
-  let originalLines = fs.readFileSync(scriptPath).toString().split(os.EOL);
-  let newLines: string[] = [];
-
-  // Modify the lines
-  for (let line of originalLines) {
-    if (line == "TEST.NEW") line = "TEST.REPLACE";
-    newLines.push(line);
-  }
-
-  // Join the modified lines back into a single string
-  const modifiedContent = newLines.join("\n");
-
-  // Write the modified content back to the file
-  fs.writeFileSync(scriptPath, modifiedContent, "utf8");
-}
-
-export async function openTestScript(nodeID: string) {
-  // this can get called for a unit, environment, function, or test
-
-  const testNode: testNodeType = getTestNode(nodeID);
-
-  const enclosingDirectory = path.dirname(testNode.enviroPath);
-  const scriptPath = testNode.enviroPath + ".tst";
-
-  let commandToRun: string = `${clicastCommandToUse} ${getClicastArgsFromTestNode(
-    testNode
-  )} test script create ${scriptPath}`;
-  const commandStatus = executeCommand(commandToRun, enclosingDirectory);
-  if (commandStatus.errorCode == 0) {
-    // Improvement needed:
-    // It would be nice if vcast generated the scripts with TEST.REPLACE, but for now
-    // convert TEST.NEW to TEST.REPLACE so doing an "immediate load" works without error
-    convertTestScriptContents(scriptPath);
-
-    // open the script file for editing
-    vscode.workspace.openTextDocument(scriptPath).then(
-      (doc: vscode.TextDocument) => {
-        vscode.window.showTextDocument(doc);
-      },
-      (error: any) => {
-        vectorMessage(error);
-      }
-    );
-  }
-}
-
-
-
-async function adjustScriptContentsBeforeLoad (scriptPath:string) {
-
-  // There are some things that need updating before we can load the 
-  // script into VectorCAST:
-  //   - The requirement key lines need to be split into two lines
-  //     We insert lines like TEST.REQUIREMENT_KEY: key | description,
-  //     but VectorCAST only allows the key, so we turn the description
-  //     into a comment.
-  //
-  //   - <might be more things to do later>
-
-  let originalLines = fs.readFileSync(scriptPath).toString().split("\n");
-  let newLines: string[] = [];
-  for (let line of originalLines) {
-    if (line.startsWith("TEST.REQUIREMENT_KEY:")) {
-      const keyLineParts = line.split("|");
-      if (keyLineParts.length == 2) {
-        newLines.push("-- Requirement Title: " + keyLineParts[1]);
-        newLines.push(keyLineParts[0].trim());
-      }
-      else {
-        newLines.push(line);
-      }
-    }
-    else {
-      newLines.push(line);
-    }
-  }
-  fs.writeFileSync (scriptPath, newLines.join("\n"), "utf8");
-}
-
-
-
-const url = require("url");
-export async function loadTestScript() {
-  // This gets called from the right-click editor context menu
-  // The convention is that the .tst file must be in the same directory
-  // as the environment, so we get the enviroName from parsing the
-  // .tst and get the working directory from its location
-
-  // this returns the environment directory name without any nesting
-
-  const activeEditor = vscode.window.activeTextEditor;
-  if (activeEditor) {
-    if (activeEditor.document.isDirty) {
-      // need to wait, otherwise we have a race condition with clicast
-      await activeEditor.document.save();
-    }
-
-    let scriptPath = url.fileURLToPath(activeEditor.document.uri.toString());
-    adjustScriptContentsBeforeLoad (scriptPath);
-
-    const enviroName = getEnviroNameFromScript(scriptPath);
-    if (enviroName) {
-      const enviroArg = `-e${enviroName}`;
-      const enviroPath = path.join(path.dirname(scriptPath), enviroName);
-      let commandToRun: string = `${clicastCommandToUse} ${enviroArg} test script run ${scriptPath}`;
-      const commandStatus = executeCommand(
-        commandToRun,
-        path.dirname(scriptPath)
-      );
-      // if the script load fails, executeCommand will open the message pane ...
-      // if the load passes, we want to give the user an indication that it worked
-      if (commandStatus.errorCode == 0) {
-        vectorMessage("Script loaded successfully ...");
-        // Maybe this will be annoying to users, but I think
-        // it's good to know when the load is complete.
-        vscode.window.showInformationMessage(`Test script loaded successfully`);
-      }
-
-      updateTestPane(enviroPath);
-    } else {
-      vscode.window.showErrorMessage(
-        `Could not determine environment name, required "-- Environment: <enviro-name> comment line is missing.`
-      );
-    }
-  }
-}
 
 function processVCtestData(
   controller: TestController,
